feat(main-page): 하단 내비게이바 추가
